==> 00.CPES_2024_set_up_file_paths.R <==
# Name of file: 00.CPES_2024_set_up_file_paths.R
# 
# Original author(s): Catriona Haddow
#   
# Written/run on: Posit Workbench - RStudio R 4.1.2
# 
# Description of content: Set up file paths to be used throughout project
# 
# Approximate run time: <1 min
# 
# Approximate memory usage: 245mib

lookup_path   <- "/conf/bss/CPES/2024/Data/Lookup/"
data_path     <- "/conf/bss/CPES/2024/Data/"
analysis_output_path <- "/conf/bss/CPES/2024/Output/analysis_output/"
output_path   <- "/conf/bss/CPES/2024/Output/"
data_path_2018 <- "/conf/bss/CPES/Reprocess_in_R/2018/Output/analysis_output/"
data_path_2015 <- "/conf/bss/CPES/Reprocess_in_R/2015/Output/analysis_output/"
